Fix CMake System and IMU interface
fix: fix CMakeLists.txt in modules

fix: fix imu_dgram_t

feat: add SPI definitions

fix: remove redundant macros

BREAKING CHANGES

refactor: use C struct instead of macro to separate interface / realization

fix: remove uart_buffer_len from imu_dgram_t, replace with buffer_delay_us

fix: remove level field from restful GET[/v1/imu/status] and replace with power
