+++ CMakeLists.txt
@@ -2,6 +2,7 @@
 # CMakeLists in this exact order for cmake to work correctly
 cmake_minimum_required(VERSION 3.5)
 
+# Add git tag to firmware version
 EXECUTE_PROCESS(COMMAND git describe --tags --abbrev=0
         TIMEOUT 5
         OUTPUT_VARIABLE GIT_VERSION
@@ -10,24 +11,31 @@ EXECUTE_PROCESS(COMMAND git describe --tags --abbrev=0
 MESSAGE(STATUS "building from git tag ${GIT_VERSION}")
 add_definitions(-DCONFIG_FIRMWARE_VERSION=\"${GIT_VERSION}\")
 
+
+# All available IMU types
+add_definitions(-DIMU_TYPE_GY95=95)
+add_definitions(-DIMU_TYPE_HI229=229)
+add_definitions(-DIMU_TYPE_BNO08X=80)
+
+# Select IMU module
+set(IMU_MODULE hi229) # hi229 or bno08x
+add_definitions(-DCONFIG_IMU_TYPE=IMU_TYPE_HI229) # IMU_TYPE_HI229 or IMU_TYPE_BNO08X, depending on the selected module
+
+
+# Introduce esp-iot-solution
 # set(IOT_SOLUTION_PATH  ${CMAKE_CURRENT_LIST_DIR}/esp-iot-solution)
 # include(${IOT_SOLUTION_PATH}/component.cmake)
-
 # set(EXTRA_COMPONENT_DIRS ${HOMEKIT_PATH}/components/homekit ${HOMEKIT_PATH}/examples/common ${IOT_SOLUTION_PATH}/components/button)
 
+# Introduce used components
 set(
         EXTRA_COMPONENT_DIRS
-        ${CMAKE_SOURCE_DIR}/lib/hi229
-        ${CMAKE_SOURCE_DIR}/lib/bno08x
-        ${CMAKE_SOURCE_DIR}/lib/battery
-        ${CMAKE_SOURCE_DIR}/lib/ring_buf
-        ${CMAKE_SOURCE_DIR}/lib/libudp
-        ${CMAKE_SOURCE_DIR}/lib/spatial
-        ${CMAKE_SOURCE_DIR}/lib/battery
+        ${CMAKE_SOURCE_DIR}/lib/${IMU_MODULE}
+        ${CMAKE_SOURCE_DIR}/lib/battery ${CMAKE_SOURCE_DIR}/lib/ring_buf ${CMAKE_SOURCE_DIR}/lib/libudp ${CMAKE_SOURCE_DIR}/lib/spatial
         ${CMAKE_SOURCE_DIR}/components
 )
 
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
+set(ENV{IMU_MODULE} "${IMU_MODULE}")
 project(rfmarkit-esp-node)
-
 #idf_build_set_property(COMPILE_OPTIONS "-O3" APPEND)
