+++ components/imu/CMakeLists.txt
@@ -1,6 +1,4 @@
-FILE(GLOB_RECURSE sources *.c)
 idf_component_register(
-        SRCS ${sources}
         INCLUDE_DIRS . ${CMAKE_SOURCE_DIR}/include
-        REQUIRES sys hi229
+        REQUIRES sys $ENV{IMU_MODULE}
 )
