+++ components/sys/CMakeLists.txt
@@ -2,5 +2,5 @@ FILE(GLOB_RECURSE sources *.c)
 idf_component_register(
         SRCS ${sources}
         INCLUDE_DIRS . ${CMAKE_SOURCE_DIR}/include
-        REQUIRES imu blink esp_https_ota nvs_flash ring_buf app_update libudp bt ble rest_controller spatial battery
+        REQUIRES blink esp_https_ota nvs_flash ring_buf app_update libudp bt ble rest_controller spatial battery
 )
