+++ lib/battery/battery.h
@@ -7,6 +7,8 @@
 
 #define CONFIG_BATTERY_EN_VALUE             0
 
+#define battery_delay_ms(x) vTaskDelay((x) / portTICK_PERIOD_MS)
+
 esp_err_t battery_msp_init(void);
 
 int battery_read_level(void);
