+++ lib/libudp/CMakeLists.txt
@@ -2,5 +2,5 @@ FILE(GLOB_RECURSE sources *.c)
 idf_component_register(
         SRCS ${sources}
         INCLUDE_DIRS .
-        REQUIRES lwip sys imu
+        REQUIRES lwip
 )
