+++ lib/spatial/CMakeLists.txt
@@ -2,5 +2,5 @@ FILE(GLOB_RECURSE sources *.c)
 idf_component_register(
         SRCS ${sources}
         INCLUDE_DIRS . ${CMAKE_SOURCE_DIR}/include
-        REQUIRES sys imu
+        REQUIRES sys
 )
